py/services/stats-aggregator/pyproject.toml
[tool.poetry]
name = "stats-aggregator"
version = "0.1.0"
description = ""
authors = ["Aleksey Dalekin <adalekin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
aiohealthcheck = "^2018.7.1"
aiokafka = "^0.7.2"
aiorun = "^2021.8.1"
cassandra-driver = "^3.25.0"
environs = "^9.3.3"
httpx = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
